+++ ErsatzTV.Application/Images/Commands/UpdateImageFolderDuration.cs
@@ -1,0 +1,3 @@
+namespace ErsatzTV.Application.Images;
+
+public record UpdateImageFolderDuration(int LibraryFolderId, int? ImageFolderDuration) : IRequest;