+++ ErsatzTV.Application/Images/Queries/GetImageFolders.cs
@@ -1,0 +1,3 @@
+namespace ErsatzTV.Application.Images;
+
+public record GetImageFolders(Option<int> LibraryFolderId) : IRequest<List<ImageFolderViewModel>>;