+++ ErsatzTV.Core/Scheduling/PlayoutModeSchedulerBase.cs
@@ -179,9 +179,9 @@ public abstract class PlayoutModeSchedulerBase<T> : IPlayoutModeScheduler<T> whe
 
     protected static TimeSpan DurationForMediaItem(MediaItem mediaItem)
     {
-        if (mediaItem is Image)
+        if (mediaItem is Image image)
         {
-            return TimeSpan.FromSeconds(Domain.Image.DefaultSeconds);
+            return TimeSpan.FromSeconds(image.ImageMetadata.Head().DurationSeconds ?? Image.DefaultSeconds);
         }
 
         MediaVersion version = mediaItem.GetHeadVersion();