+++ CHANGELOG.md
@@ -23,8 +23,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - The templates contain comments describing which fields are available for use in the templates
 - Add *experimental* and *incomplete* `Images` library kind
   - Image libraries have fallback metadata added like Other Video libraries (every folder is a tag)
-  - Image library items currently *all* have a duration of 15 seconds
-    - Future updates will allow custom/distinct durations
+  - Image library items currently default to a duration of 15 seconds
+    - The `Media` > `Images` page can be used to configure image durations at a folder level
+    - Child folders with unset durations will inherit the closest ancestor's duration
 
 ### Fixed
 - Fix antiforgery error caused by reusing existing browser tabs across docker container restarts